requirements.txt
matplotlib
pandas
yfinance
pillow